pyproject.toml
[tool.poetry]
name = "drfproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
black = "^24.4.2"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
isort = "^5.13.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
